# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/user-ms/v1/customer/users/self": {
            "get": {
                "description": "This endpoint allows current login user fetch his/her profile in JSON format.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get User Profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.UserProfileVO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "This endpoint allows current login user update his/her profile in JSON format.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update User Profile",
                "parameters": [
                    {
                        "description": "User profile to update",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.UserProfileVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.UserProfileVO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    }
                }
            }
        },
        "/user-ms/v1/customer/users/self/addresses": {
            "get": {
                "description": "This endpoint allows user list all their addresses.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserAddress"
                ],
                "summary": "List User Addresses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.UserAddressVO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "This endpoint allows user create new address in JSON format.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserAddress"
                ],
                "summary": "Create New User Address",
                "parameters": [
                    {
                        "description": "user address to create",
                        "name": "address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.UserAddressVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.UserAddressVO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    }
                }
            }
        },
        "/user-ms/v1/customer/users/self/addresses/{address_id}": {
            "put": {
                "description": "This endpoint allows user update their old address in JSON format.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserAddress"
                ],
                "summary": "Update existing User Address",
                "parameters": [
                    {
                        "description": "user address to update",
                        "name": "address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.UserAddressVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.UserAddressVO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "This endpoint allows user delete their address.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserAddress"
                ],
                "summary": "Delete User Address",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Address ID",
                        "name": "address_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    }
                }
            }
        },
        "/user-ms/v1/{client}/login": {
            "post": {
                "description": "Authenticates a user with their email and password and returns a token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "User Login",
                "parameters": [
                    {
                        "description": "User login information",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.UserLoginVO"
                        }
                    },
                    {
                        "enum": [
                            "customer",
                            "merchant"
                        ],
                        "type": "string",
                        "description": "Client identifier",
                        "name": "client",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful, returns auth token in cookie",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user-ms/v1/{client}/logout": {
            "post": {
                "description": "invalidates the user's auth token cookie.",
                "tags": [
                    "Authentication"
                ],
                "summary": "User Logout",
                "parameters": [
                    {
                        "enum": [
                            "customer",
                            "merchant"
                        ],
                        "type": "string",
                        "description": "Client identifier",
                        "name": "client",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/user-ms/v1/{client}/users": {
            "post": {
                "description": "This endpoint allows a new user to register by providing their details in JSON format.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Register"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration details",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.UserLoginVO"
                        }
                    },
                    {
                        "enum": [
                            "customer",
                            "merchant"
                        ],
                        "type": "string",
                        "description": "Client identifier",
                        "name": "client",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    }
                }
            }
        },
        "/user-ms/v1/{client}/users/activate": {
            "put": {
                "description": "This endpoint allows a new user to activate by providing their verification code in JSON format.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Register"
                ],
                "summary": "Activate a new user",
                "parameters": [
                    {
                        "description": "User activate request",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.UserActivateReq"
                        }
                    },
                    {
                        "enum": [
                            "customer",
                            "merchant"
                        ],
                        "type": "string",
                        "description": "Client identifier",
                        "name": "client",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "data.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "err_msg": {
                    "type": "string"
                }
            }
        },
        "data.UserActivateReq": {
            "type": "object",
            "required": [
                "code"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 6,
                    "minLength": 6
                }
            }
        },
        "data.UserAddressVO": {
            "type": "object",
            "required": [
                "contact_phone",
                "country",
                "detail",
                "first_name",
                "last_name",
                "zip_code"
            ],
            "properties": {
                "city": {
                    "type": "string"
                },
                "contact_phone": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "detail": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_default": {
                    "type": "boolean"
                },
                "last_name": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                },
                "zip_code": {
                    "type": "string"
                }
            }
        },
        "data.UserLoginVO": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "data.UserProfileVO": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "default_address": {
                    "$ref": "#/definitions/data.UserAddressVO"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}